Dynamic proxy for EObjects that are not native FacadeObjects.
Also re-casts the IXMIIDCopier extension protocol as more
generally ICopier with responsibility for creating the
copy in the merge result in the first place.  This lets
a façade-specific copier create the adapters for models that
do not natively implement the FacadeObject protocol, to give
EMF Compare access to the expected eResource() and
eDirectResource() semantics.

Includes re-generation of the J2EE test façade model
without the FacadeObject root interface (as though
it were a closed third-party model).

Change-Id: I1ee144c1500eb1f17a9796614c6af6e01e561ae4