Ensure consistency of EReferences in façades.
Ensure that accessors for EReference features in
façades that are dynamic proxies present a consistent
dynamic-proxy view of related objects.  Similarly
for reflect access to containers and contents and
for details of notifications.

Likewise ensure that proxies passed in to EReference
mutators and EOperations are unwrapped as necessary
for the internal workings of the EMF-generated code.

Change-Id: I36aa1be0decdad85fbd2ddfacf6564480fc1594a