Pluggable computation of extrinsic IDs for façade objects.
Fix the performance of compare/merge for the opaque expressions
façade, which was relying on the much more expensive structural
matching of expression body objects because they do not trace
to distinct elements in the underlying UML model with their own
identifiers.

This adds a new API to the ICopier protocol, to obtain an
identifier for the object.  This seems a sensible owner of that
responsibility as it is closely related to copying XMI IDs.

Change-Id: I64ed83b7c0acd4d527140c1df197d5cf31de861b